# manga-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: Mangarr Manga-Gateway API
  version: "1.0.0"
  description: >
    Single external "manga-gateway" process exposing BOTH a search/indexer surface
    (the Jackett/Torznab analog) and a download surface (the SABnzbd/qBittorrent analog).
    One authenticated secure-site session + one anti-bot solver are shared across both
    surfaces. Mangarr registers this one host as one IIndexer and one IDownloadClient.

    Conceptual model: Torznab (caps -> gated search -> normalized release list).
    Wire format: JSON over HTTP(S). See spikes 001 (download), 002 (search), 003 (Mangarr migration).
servers:
  - url: "http://localhost:9191/api/v1"
    description: Default localhost bind (recommended). Prepend UrlBase if configured.
security:
  - ApiKeyHeader: []
  - ApiKeyQuery: []
paths:
...
  /caps:
    get:
      summary: Capabilities handshake — sources, supported params, limits, formats
      operationId: getCaps
      responses:
        "200":
          description: Capabilities document (clients should cache 6-12h)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Capabilities" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /recent:
    get:
      summary: Recent-release feed (RSS-sync peer of IIndexer.FetchRecent)
      operationId: getRecent
      parameters:
        - { name: sources,   in: query, schema: { type: string }, description: "CSV of source keys; omit = all enabled" }
        - { name: languages, in: query, schema: { type: string }, description: "CSV of BCP-47 codes" }
        - { name: limit,     in: query, schema: { type: integer, default: 100, maximum: 100 } }
        - { name: since,     in: query, schema: { type: string, format: date-time }, description: "Return items newer than this (optimization)" }
      responses:
        "200":
          description: Releases, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReleaseListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ─────────────────────────── DOWNLOAD SURFACE (spike 001) ───────────────────────────
...
components:
  securitySchemes:
    ApiKeyHeader: { type: apiKey, in: header, name: X-Api-Key }
    ApiKeyQuery:  { type: apiKey, in: query,  name: apikey }
  responses:
    Unauthorized:
      description: Missing/invalid API key
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, enum: [auth, rate_limited, source_unavailable, bad_request, internal] }
            message: { type: string }
    Capabilities:
      type: object
      required: [gatewayVersion, sources, supportedSearchParams, limits, downloadFormats]
      properties:
        gatewayVersion: { type: string }
        sources:
          type: array
          items: { $ref: "#/components/schemas/SourceCap" }
        supportedSearchParams:
          type: array
          items: { type: string }
          example: [q, mangadexId, anilistId, malId, chapter, volume, language, sourceKey]
        limits:
          type: object
          properties:
            defaultPageSize: { type: integer, example: 50 }
            maxPageSize:     { type: integer, example: 100 }
        downloadFormats:
          type: array
          items: { type: string, enum: [cbz, cbt, folder] }
    SourceCap:
      type: object
      required: [key, name, enabled, supportsSearch, supportsRecent]
      properties:
        key:                { type: string, example: "comix.to" }
        name:               { type: string, example: "Comix" }
        enabled:            { type: boolean }
        supportsSearch:     { type: boolean }
        supportsRecent:     { type: boolean }
        idTypes:            { type: array, items: { type: string }, example: [mangadexId] }
        languages:          { type: array, items: { type: string }, example: [en, es-la] }
        rateLimitPerMinute: { type: integer }
        antibot:            { type: string, enum: [none, cloudflare, "cloudflare+encrypted"] }
...
    ReleaseListResponse:
      type: object
      required: [releases]
      properties:
        releases:
          type: array
          items: { $ref: "#/components/schemas/Release" }
        warnings:
          type: array
          items: { $ref: "#/components/schemas/SourceWarning" }
    SourceWarning:
      type: object
      required: [sourceKey, code, message]
      properties:
        sourceKey: { type: string }
        code:      { type: string, example: source_degraded }
        message:   { type: string }
    Release:
      type: object
      required: [guid, title, sourceKey, downloadHandle, publishDate]
      properties:
        guid:            { type: string, description: "dedup key; Mangarr de-dupes by this" }
        title:           { type: string, description: "MUST be parseable by Mangarr's MangaParser (name + chapter + group/lang)" }
        sourceKey:       { type: string }
        downloadHandle:  { type: string, description: "opaque; submitted back to POST /downloads. TTL >= 30 min." }
        infoUrl:         { type: string, nullable: true }
        publishDate:     { type: string, format: date-time }
        mangaTitle:      { type: string, nullable: true, description: "Structured series title. Fully populated when known; the intended primary path for a future Mangarr (parsing title is then optional). null only when genuinely unknown." }
        chapterNumber:   { type: number, nullable: true, description: "Decimal chapter number, first-class data. Fully populated when known (>=3 decimal places preserved) and the intended primary path for a future Mangarr; parsing the title becomes optional. null only when genuinely unknown." }
        volume:          { type: integer, nullable: true, description: "Structured volume number. Populated when known; first-class data alongside chapterNumber. null only when genuinely unknown." }
        language:        { type: string, nullable: true, description: "BCP-47 translated language -> ReleaseInfo.TranslatedLanguage. First-class data; fully populated when known. null only when genuinely unknown." }
        scanlationGroup: { type: string, nullable: true, description: "Structured scanlation/translation group. Populated when known; first-class data the future Mangarr reads directly. null only when genuinely unknown." }
        pageCount:       { type: integer, nullable: true, description: "Page count for the chapter. Populated when known; first-class structured data. null only when genuinely unknown." }
        sizeBytes:       { type: integer, default: 0 }
        ids:             { type: object, additionalProperties: true, nullable: true, description: "Cross-reference source ids (e.g. mangadexChapterId/mangadexMangaId). First-class structured data; populated when known." }
...